+++ react/hooks/useToolbar.ts
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import { useIntl } from 'react-intl'
 import { Item } from 'vtex.checkout-graphql'
 
+import { useCheckoutB2BContext } from '../CheckoutB2BContext'
 import type { WithToast } from '../typings'
 import { messages } from '../utils'
 import { useOrderFormCustom } from './useOrderFormCustom'
@@ -10,6 +11,7 @@ import { usePlaceOrder } from './usePlaceOrder'
 export function useToolbar(showToast: WithToast['showToast']) {
   const { formatMessage } = useIntl()
   const { orderForm } = useOrderFormCustom()
+  const { pending } = useCheckoutB2BContext()
   const { placeOrder, isLoading, isSuccess } = usePlaceOrder(showToast)
 
   const [searchTerm, setSearchTerm] = useState<string>('')
@@ -49,7 +51,7 @@ export function useToolbar(showToast: WithToast['showToast']) {
       onSubmit: handleSubmit,
     },
     newLine: {
-      disabled: isLoading || isSuccess,
+      disabled: isLoading || isSuccess || pending,
       isLoading,
       label: formatMessage(messages.placeOrder),
       handleCallback: placeOrder,